+++ scripts/classes/metamagic-careful-spell.js
@@ -3,13 +3,20 @@
 	from the spell's full force. To do so, you spend 1 sorcery point and choose a number of those creatures up to your
 	Charisma modifier (minimum of one creature). A chosen creature automatically succeeds on its saving throw against the spell.
 */
-const version = "11.0";
+const version = "11.1";
 const optionName = "Careful Spell";
 const baseName = "Font of Magic";
 const cost = 1;
 
 try {
 	if (args[0].macroPass === "preItemRoll") {
+		let usesLeft = HomebrewHelpers.getAvailableSorceryPoints(actor);
+		if (!usesLeft || usesLeft < cost) {
+			console.error(`${optionName} - not enough Sorcery Points left`);
+			ui.notifications.error(`${optionName} - not enough Sorcery Points left`);
+			return false;
+		}
+
 		// check max targets
 		let maxTargets = Math.max(actor.system.abilities.cha.mod, 1);
 		if (workflow.preSelectedTargets.length > maxTargets) {
@@ -17,27 +24,12 @@ try {
 			return false;
 		}
 
-		let fontOfMagic = actor.items.find(i => i.name === optionName);
-		if (fontOfMagic) {
-			let usesLeft = fontOfMagic.system.uses?.value ?? 0;
-			if (!usesLeft || usesLeft < cost) {
-				console.error(`${optionName} - not enough Sorcery Points left`);
-				ui.notifications.error(`${optionName} - not enough Sorcery Points left`);
-			}
-			else {
-				const newValue = fontOfMagic.system.uses.value - cost;
-				await fontOfMagic.update({"system.uses.value": newValue});
-				return true;
-			}
-		}
-		else {
-			console.error(`${optionName} - no ${baseName} item on actor`);
-			ui.notifications.error(`${optionName} - no ${baseName} item on actor`);
-		}
-
-		return false;
+		return true;
+	}
+	else if (args[0].macroPass === "postActiveEffects") {
+		await HomebrewHelpers.reduceAvailableSorceryPoints(actor, cost)
 	}
 	
 } catch (err)  {
-    console.error(`Metamagic: ${optionName} ${version}`, err);
+	console.error(`${optionName}: ${version}`, err);
 }