+++ scripts/classes/metamagic-subtle-spell.js
@@ -1,26 +1,18 @@
 /*
 	When you cast a spell, you can spend 1 sorcery point to cast it without any somatic or verbal components.
 */
-const version = "11.0";
+const version = "11.1";
 const optionName = "Metamagic: Subtle Spell";
-const baseName = "Font of Magic";
 const cost = 1;
-const mutationFlag = "subtle-spell-item";
+const flagName = "subtle-spell-item";
+const _flagGroup = "fvtt-trazzm-homebrew-5e";
 
 try {
 	if (args[0].macroPass === "preItemRoll") {
-		let fontOfMagic = actor.items.find(i => i.name === optionName);
-		if (fontOfMagic) {
-			let usesLeft = fontOfMagic.system.uses?.value ?? 0;
-			if (!usesLeft || usesLeft < cost) {
-				console.error(`${optionName} - not enough Sorcery Points left`);
-				ui.notifications.error(`${optionName} - not enough Sorcery Points left`);
-				return false;
-			}
-		}
-		else {
-			console.error(`${optionName} - no ${baseName} item on actor`);
-			ui.notifications.error(`${optionName} - no ${baseName} item on actor`);
+		let usesLeft = HomebrewHelpers.getAvailableSorceryPoints(actor);
+		if (!usesLeft || usesLeft < cost) {
+			console.error(`${optionName} - not enough Sorcery Points left`);
+			ui.notifications.error(`${optionName} - not enough Sorcery Points left`);
 			return false;
 		}
 
@@ -95,8 +87,8 @@ try {
 						await warpgate.mutate(token.document, updates, {}, { name: itemName });
 												
 						// track target info on the actor
-						DAE.setFlag(actor, mutationFlag, {itemName : itemName } );
-						await consumeResource(actor, resKey, cost);
+						await HomebrewHelpers.reduceAvailableSorceryPoints(actor, cost)
+						await actor.setFlag(_flagGroup, flagName, {itemName : itemName });
 					}
 				},
 				Cancel:
@@ -108,11 +100,11 @@ try {
 
 	}
 	else if (args[0] === "off") {
-		let flag = DAE.getFlag(actor, mutationFlag);
+		let flag = actor.getFlag(_flagGroup, flagName);
 		if (flag) {
+			await actor.unsetFlag(_flagGroup, flagName);
 			const itemName = flag.itemName;
-			let restore = await warpgate.revert(token.document, itemName);
-			DAE.unsetFlag(actor, mutationFlag);
+			await warpgate.revert(token.document, itemName);
 		}
 	}