+++ scripts/items/drow-poison.js
@@ -1,4 +1,4 @@
-/*
+drow-poison.js/*
 	This poison is typically made only by the drow, and only in a place far removed from sunlight. A creature subjected to this poison must succeed on a DC 13 Constitution saving throw or be Poisoned for 1 hour. If the saving throw fails by 5 or more, the creature is also Unconscious while poisoned in this way. The creature wakes up if it takes damage or if another creature takes an action to shake it awake.
 */
 const version = "11.2";